requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
psycopg==3.0.8
psycopg-binary==3.0.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
